+++ castor.php
@@ -5,6 +5,7 @@ use Castor\Attribute\AsTask;
 use function Castor\import;
 
 import(__DIR__ . '/examples');
+import(__DIR__ . '/tools/phar/castor');
 import(__DIR__ . '/tools/watcher/castor');
 
 #[AsTask(description: 'hello')]