+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add support for better handling of option without value
 * Fix the stubs generation when castor is installed via composer
 * Fix the initial castor.php file generated for new projects
+* Fix watch() function
 
 ## 0.1.0 (2023-05-21)